Add jwt to endpoints

+++ app.py
@@ -1,5 +1,6 @@
 import os
 from datetime import datetime, timezone, timedelta
+from functools import wraps
 from typing import TypedDict
 
 import bcrypt
@@ -18,6 +19,7 @@ class User(TypedDict):
 class Task(TypedDict):
     title: str
     description: str
+    user_id: ObjectId
 
 
 app = Flask(__name__)
@@ -33,6 +35,32 @@ client = MongoClient(uri)
 database = client.get_database('todo')
 users_collection = database.get_collection('users')
 tasks_collection = database.get_collection('tasks')
+
+
+def token_required(f):
+    @wraps(f)
+    def decorator(*args, **kwargs):
+        token = None
+
+        if 'Authorization' in request.headers:
+            token = request.headers['Authorization'].split(" ")[1]
+
+        if not token:
+            return jsonify({'message': 'token is missing'}), 401
+
+        try:
+            decoded_token = jwt.decode(token, SECRET_KEY, algorithms=[JWT_ALGORITHM])
+            current_user = users_collection.find_one({'_id': ObjectId(decoded_token['user_id'])})
+            if not current_user:
+                return jsonify({'message': 'user not found'}), 404
+        except jwt.ExpiredSignatureError:
+            return jsonify({'message': 'token has expired'}), 401
+        except jwt.InvalidTokenError:
+            return jsonify({'message': 'invalid token'}), 401
+
+        return f(current_user, *args, **kwargs)
+
+    return decorator
 
 
 @app.route('/status', methods=['GET'])
@@ -91,25 +119,26 @@ def login():
 
 
 @app.route('/todos', methods=['POST'])
-def create_task():
+@token_required
+def create_task(current_user):
     task_data = request.json
     tasks_collection.insert_one(
         Task(
             title=task_data['title'],
-            description=task_data['description']
+            description=task_data['description'],
+            user_id=current_user['_id']
         )
     )
     return jsonify({'message': 'task created successfully'}), 201
 
 
 @app.route('/todos', methods=['GET'])
-def get_tasks():
-    tasks = tasks_collection.find()
+@token_required
+def get_tasks(current_user):
+    user_object_id = ObjectId(current_user['_id'])
 
-    tasks_list = [
-        {**task, '_id': str(task['_id'])}
-        for task in tasks
-    ]
+    tasks = tasks_collection.find({'user_id': user_object_id})
+    tasks_list = [{**task, '_id': str(task['_id']), 'user_id': str(task['user_id'])} for task in tasks]
 
     return jsonify({
         'message': 'tasks retrieved successfully',
@@ -118,18 +147,28 @@ def get_tasks():
 
 
 @app.route('/todos/<task_id>', methods=['PUT'])
-def update_task(task_id):
+@token_required
+def update_task(current_user, task_id):
     task_object_id = ObjectId(task_id)
-    updated_task_data = request.json
 
+    task = tasks_collection.find_one({'_id': task_object_id})
+    if task['user_id'] != ObjectId(current_user['_id']):
+        return jsonify({'message': 'unauthorized to update this task'}), 403
+
+    updated_task_data = request.json
     tasks_collection.update_one({'_id': task_object_id}, {'$set': updated_task_data})
 
     return jsonify({'message': 'task updated successfully'}), 200
 
 
 @app.route('/todos/<task_id>', methods=['DELETE'])
-def delete_task(task_id):
+@token_required
+def delete_task(current_user, task_id):
     task_object_id = ObjectId(task_id)
+
+    task = tasks_collection.find_one({'_id': task_object_id})
+    if task['user_id'] != ObjectId(current_user['_id']):
+        return jsonify({'message': 'unauthorized to delete this task'}), 403
 
     tasks_collection.delete_one({'_id': task_object_id})
